[FEAT] Photo Manager 따로 생성(#545)
