+++ Manito/Manito/Presentation/Error/LetterScene/LetterImageError.swift
@@ -10,7 +10,6 @@ import Foundation
 enum LetterImageError: LocalizedError {
     case invalidImage
     case invalidPhotoLibrary
-    case cantloadPhoto
 }
 
 extension LetterImageError {
@@ -18,7 +17,6 @@ extension LetterImageError {
         switch self {
         case .invalidImage: return TextLiteral.letterImageViewControllerErrorMessage
         case .invalidPhotoLibrary: return TextLiteral.letterImageViewControllerErrorMessage
-        case .cantloadPhoto: return TextLiteral.letterPhotoViewFail
         }
     }
 }
